requirements.txt
ipdb==0.10.1
ipython==5.1.0
pyperclip==1.5.27
PyHamcrest==1.9.0
pytest==3.0.3
pytest-xdist==1.15.0
python-pom==1.0.2
selenium==2.53.6
Sphinx==1.4.8
xvfbwrapper==0.2.8
